backend/app/auth.py
import bcrypt
from jose import jwt, JWTError
from datetime import datetime, timedelta

SECRET_KEY = "your_secret_key_change_this"
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_HOURS = 8


# Password hashing
def hash_password(password: str):
    return bcrypt.hashpw(password.encode(), bcrypt.gensalt()).decode()


def verify_password(password: str, hashed: str):
    return bcrypt.checkpw(password.encode(), hashed.encode())


# JWT
def create_token(admin_id: int):
    payload = {
        "admin_id": admin_id,
        "exp": datetime.utcnow() + timedelta(hours=ACCESS_TOKEN_EXPIRE_HOURS)
    }
    return jwt.encode(payload, SECRET_KEY, algorithm=ALGORITHM)


def verify_token(token: str):
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        return payload["admin_id"]
    except JWTError:
        return None