backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
joblib
numpy
pandas
catboost
requests
bcrypt
python-jose